+++ packages/block-editor/src/components/list-view/use-list-view-collapse-items.js
@@ -10,7 +10,7 @@ import { useSelect } from '@wordpress/data';
 import { store as blockEditorStore } from '../../store';
 import { unlock } from '../../lock-unlock';
 
-export default function useListViewCollapseItems( { collapseAll, expand } ) {
+export default function useListViewCollapseItems( { updateExpansion } ) {
 	const { expandedBlock, getBlockParents } = useSelect( ( select ) => {
 		const { getBlockParents: _getBlockParents, getExpandedBlock } = unlock(
 			select( blockEditorStore )
@@ -26,8 +26,8 @@ export default function useListViewCollapseItems( { collapseAll, expand } ) {
 		if ( expandedBlock ) {
 			const blockParents = getBlockParents( expandedBlock, false );
 			// Collapse all blocks and expand the block's parents.
-			collapseAll();
-			expand( blockParents );
+			updateExpansion( { type: 'clear' } );
+			updateExpansion( { type: 'expand', clientIds: blockParents } );
 		}
-	}, [ collapseAll, expand, expandedBlock, getBlockParents ] );
+	}, [ updateExpansion, expandedBlock, getBlockParents ] );
 }