+++ test/e2e/specs/interactivity/tovdom-islands.spec.ts
@@ -8,9 +8,11 @@ test.describe( 'toVdom - islands', () => {
 		await utils.activatePlugins();
 		await utils.addPostWithBlock( 'test/tovdom-islands' );
 	} );
+
 	test.beforeEach( async ( { interactivityUtils: utils, page } ) => {
 		await page.goto( utils.getLink( 'test/tovdom-islands' ) );
 	} );
+
 	test.afterAll( async ( { interactivityUtils: utils } ) => {
 		await utils.deactivatePlugins();
 		await utils.deleteAllPosts();